requirements.txt
websocket-client
rel
